\name{GetBack}
\alias{GetBack}
\title{Attempts to get a result package back based on the informations stored in task package object.}
\usage{
GetBack(task_package_info, interface, simplified_output = TRUE)
}
\arguments{
  \item{task_package_info}{the output, three-element list of the \code{SendOut} function}
  \item{interface}{an interface object generated by one of the \code{generate_interface_*} functions.}
  \item{simplified_output}{default is \code{TRUE}; if set to \code{FALSE}, a list is returned containing both the input expression and the output object after evaluation.}
}

\description{The function checks the presence of result packages in the common result directory. If the specific result object is existing, the function fetches and loads the result into the R session.
}

\value{
The output of the evaluated expression. If \code(simplified_output = FALSE), then a list is returned containing both the input expression and the output object after evaluation.
}

\examples{
# in the following example ifobj_gdrive is an interface generated by generate_interface_gdrive

# a simple use-case - simplified output
taskobj <- PackIn(4+2)
taskobj_sent <- SendOut(task_pack = taskobj, interface = ifobj_gdrive)
GetBack(taskobj_sent,)

> 6

# a simple use-case - non-simplified output
GetBack(taskobj_sent, interface = ifobj_gdrive, simplified_output = F)

[[1]]
$src
[1] "4 + 2"

attr(,"class")
[1] "source"

[[2]]
[1] 6

}
