\name{RunServer}
\alias{RunServer}
\title{Initiates a remete server session.}
\usage{
RunServer(interface)
}
\arguments{
  \item{interface}{an interface object generated by one of the \code{generate_interface_*} functions.}
}

\description{The function starts a server session, which periodically checks the presence of task packages in the common tasks directory, defined in the interface function. If an unevaluated task object is present, the function fetches it and runs the task.
}

\value{
The function does not return any internal objects. Messages about current processes are shown.
}

\examples{
# in the following example ifobj_gdrive is an interface generated by generate_interface_gdrive

# running an example server
RunServer(ifobj_gdrive) # starts the server
}
