\name{SendOut}
\alias{SendOut}
\title{Sends a task package to a remote server through an interface.}
\usage{
SendOut(task_pack, interface)
}
\arguments{
  \item{task_pack}{the output, two-element vector of the \code{PackIn} function}
  \item{interface}{an interface object generated by one of the \code{generate_interface_*} functions.}
}

\description{The function sends a task package to the server reachable through the interface.
}

\value{
A list of three elements; the first indicating the unique identifier of the task, the second showing the path to the task package file (RData) on the local computer. The third element is the interface function itself.
}

\examples{
# in the following example ifobj_gdrive is an interface generated by generate_interface_gdrive

# a simple use-case
taskobj <- PackIn(4+2)
taskobj_sent <- SendOut(task_pack = taskobj, interface = ifobj_gdrive)

# a more complex use-case, applying piping
taskobj <- emhace.tools::RunNetMHCpan_ng(alleles = "HLA-A02:01", peptides = peps_unq, version_number = "4.1",
                                         software_path = "/home/lhgergo/Programok/netMHCpan-4.1/netMHCpan") |>
  PackIn(objects = "peps_unq", libraries = c("purrr", "emhace.tools", "magrittr")) %>%
  SendOut(ifobj_gdrive)
}
